Replaced `ThunderstoreClientTests` with `HttpApiClientTests`
